requirements.txt
xlrd
openpyxl
pandas
yfinance
pykrx
tqdm
PyYAML
catboost
scikit-learn